[editor] bake in basic setup instead
