+++ packages/editor/lib/editor.js
@@ -11,10 +11,9 @@
  * @property {boolean} [dark=false]
  * @property {OnChange} [onChange=() => {}]
  */
-import { EditorView } from '@codemirror/view'
 import { Compartment, EditorState } from '@codemirror/state'
-import { org } from '@orgajs/cm-lang'
-import theme from './theme.js'
+import { EditorView } from '@codemirror/view'
+import { setup } from './setup.js'
 
 /**
  * @param {Config} config
@@ -31,9 +30,8 @@ export function makeEditor(config) {
   const state = EditorState.create({
     doc: content,
     extensions: [
-      org(),
+      setup,
       themeConfig.of(EditorView.theme({}, { dark })),
-      theme,
       extensions,
     ],
   })
